Split DependencyException into InvalidArgumentException & NotFoundException
