+++ src/Framework/ClassResolver/DependencyResolver/DependencyInvalidArgumentException.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Gacela\Framework\ClassResolver\DependencyResolver;
 
-use RuntimeException;
+use InvalidArgumentException;
 
-final class DependencyException extends RuntimeException
+final class DependencyInvalidArgumentException extends InvalidArgumentException
 {
     public static function noParameterTypeFor(string $parameter): self
     {
@@ -16,14 +16,5 @@ final class DependencyException extends RuntimeException
     public static function unableToResolve(string $parameter, string $className): self
     {
         return new self("Unable to resolve [{$parameter}] from {$className}");
-    }
-
-    public static function mapNotFoundForClassName(string $className): self
-    {
-        $message = <<<TXT
-Did you forget to map this interface to a concrete class? No concrete class was found that implements:
-{$className}
-TXT;
-        return new self($message);
     }
 }
